Support multiple namespaces in type hierarchy
Since a type in one model can extend a type from a different model, the typeHierarchy needs to return the namespace and name. In the FHIR model, this most often happens with FHIR types extending System.Any.  This commit doesn't fully fix support for type hierarchy w/ System types, but it partially supports and fakes it for now.