CQL Exec FHIR 1.5.0 - Is Support (#15)
Compatible w/ CQL Execution 2.0.0 and provides necessary support for CQL Execution's `is` and `as` operator implementations.  Squashed from the following commits:

* Initial pass at an _is function, which takes in a typeSpec from CQL-execution and returns true/false, depending on whether the object (or any of its parent classes) match the requested namespace/class name

* add a _typeHierarchy method, and use it in _is

* Cleanups to _is

* Support multiple namespaces in type hierarchy

Since a type in one model can extend a type from a different model, the typeHierarchy needs to return the namespace and name. In the FHIR model, this most often happens with FHIR types extending System.Any.  This commit doesn't fully fix support for type hierarchy w/ System types, but it partially supports and fakes it for now.

* Fix for namespace not being the url from the modelInfo in _typeHierarchy results

* Moved the r4 tests into r401 as well, which verifies the fix

* Fix _typeHierarchy for DSTU2 and add DSTU2/STU3 tests

* Update dependencies to resolve security vulnerabilities

* Bump travis to Node 10 since eslint requires it

* Refactor _is and typeHierarchy to use type specifiers, like come out of the CQL, and update the tests (#13)

* v1.5.0-beta.1

Co-authored-by: Michael O'Keefe <mokeefe@mitre.org>
Co-authored-by: Chris Hossenlopp <hossenlopp@mitre.org>