Fix handling of choice elements that specify SimpleQuantity as a choice type (#46)
FHIR serializes choice elements (e.g., value[x]) by replacing the [x] with the datatype name (e.g., valueString). When cql-exec-fhir needs to process a choice element, it references the model info to see the available choices and uses their names to derive the expected property names in the JSON. This mainly works except in one case: choices that allow for a SimpleQuantity need to use the datatype name Quantity because SimpleQuantity is actually a profile of the Quantity type (and profiles are serialized using their base type name). This commit fixes processing of choices by adding a function to normalize choice type names that are used in FHIR JSON.

Fixes fqm-execution#352.